+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-- No items yet.
+- Error message for invalid file types
 
 ## v0.1
 - (**BREAKING**) Initial release with support for IASI and ASCAT formats.